backend/requirements.txt
fastapi
uvicorn
numpy
pandas
pillow